file rename and update package
