+++ package.json
@@ -582,7 +582,7 @@
     "dependencies": {
         "@azure/arm-authorization": "^9.0.0",
         "@azure/arm-containerregistry": "^10.1.0",
-        "@azure/arm-containerservice": "^21.1.0-beta.1",
+        "@azure/arm-containerservice": "21.2.0-beta.1",
         "@azure/arm-features": "^3.1.0",
         "@azure/arm-monitor": "^7.0.0",
         "@azure/arm-msi": "^2.1.0",
