+++ src/plugins/getPlugins.ts
@@ -1,5 +1,5 @@
 import { IActionContext } from "@microsoft/vscode-azext-utils";
-import { createAKSClusterPlugin } from "./creakteAKS/createAKSClusterPlugin";
+import { createAKSClusterPlugin } from "./createAKS/createAKSClusterPlugin";
 import { GetPluginsCommandResult } from "../types/@azure/AzureAgent";
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
